+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Breaking changes:
 - Removed ActAs, Parameters, Properties, ServiceOverrides obsolete methods from component registration. Also removed indexer, AddComponent* and Resolve from DefaultKernel and WindsorContainer (@fir3pho3nixx, #338)
 
 New Features:
+- Created Castle.Facilities.AspNetCore facility to support Mvc web applications on dotnet core and desktop clr. (@fir3pho3nixx, #120)
 - Created Castle.Facilities.AspNet.Mvc facility to support Mvc web applications on desktop clr. (@fir3pho3nixx, #283)
 - Created Castle.Facilities.AspNet.WebApi facility to support WebApi web hosted and self hosted applications on desktop clr. (@fir3pho3nixx, #283)